LayoutTests/fast/shadow-dom/activeElement-for-focused-element-in-another-shadow-expected.txt


PASS ShadowRoot.activeElement must return null the focused element is in another shadow tree when both shadow roots are open.
PASS ShadowRoot.activeElement must return null the focused element is in another shadow tree when both shadow roots are closed.
